Class Figure Calculation perimeter and area sout sum all perimeter and area

+++ src/main/java/com/nc/autumn2020/solutions/lesson5/Figure.java
@@ -2,8 +2,8 @@ package com.nc.autumn2020.solutions.lesson5;
 
 abstract class Figure {
 
-    public abstract float getPerimeter();
+    public abstract int getPerimeter();
 
-    public abstract float getArea();
+    public abstract int getArea();
 }
